Changed CleanXfac/CleanYfac back to int. Having them as floats was a constant source of problems with the ZDoom base. Fixed: Medals were not drawn at the proper position (fixes 2326). Fixed: In case "con_scaletext 0", the scoreboard entries were drawn too far left. (fixes 2403).

+++ src/m_menu.cpp
@@ -1375,8 +1375,7 @@ static void M_DrawSaveLoadCommon ()
 						listboxLeft+1, listboxTop+rowHeight*i+CleanYfac, savegamestring,
 						DTA_CleanNoMove, true, TAG_DONE);
 					screen->DrawText (SmallFont, CR_WHITE,
-						// [BB] Using Skulltag's increased precision of CleanXfac is wrong here.
-						listboxLeft+1+SmallFont->StringWidth (savegamestring)*static_cast<int>(CleanXfac),
+						listboxLeft+1+SmallFont->StringWidth (savegamestring)*CleanXfac,
 						listboxTop+rowHeight*i+CleanYfac, underscore,
 						DTA_CleanNoMove, true, TAG_DONE);
 				}
